+++ package.json
@@ -9,10 +9,8 @@
                 "@babel/core": "7.26.8",
                 "@babel/runtime": "7.26.7",
                 "@fluentui-contrib/react-resize-handle": "^0.6.1",
-                "@fluentui/react": "8.122.9",
                 "@fluentui/react-components": "9.58.3",
                 "@fluentui/react-icons": "2.0.274",
-                "@fluentui/react-icons-mdl2": "1.3.80",
                 "@microsoft/applicationinsights-react-js": "17.3.4",
                 "@microsoft/applicationinsights-web": "3.3.4",
                 "@microsoft/microsoft-graph-client": "3.0.7",
