+++ src/app/views/common/lazy-loader/component-registry/index.tsx
@@ -1,6 +1,4 @@
-import LazyResponseHeadersV9 from '../../../query-response/headers/ResponseHeadersV9';
-import LazySnippetsV9 from '../../../query-response/snippets/SnippetsV9';
-import { default as LazyCopyButtonV9 } from '../../copy-button/CopyButton';
+import LazySnippets from '../../../query-response/snippets/Snippets';
 import LazySpecificPermissions from  '../../../query-runner/request/permissions';
 import LazyStatusMessages from  '../../../app-sections/StatusMessages';
 import LazyResponseHeaders from  '../../../query-response/headers/ResponseHeaders';
@@ -27,20 +25,16 @@ export const ResponseHeaders = (props?: any) => {
   return <LazyResponseHeaders {...props} />;
 };
 
-export const ResponseHeadersV9 = (props?: any) => {
-  return <LazyResponseHeadersV9 {...props} />;
+export const Snippets = (props?: any) => {
+  return <LazySnippets {...props} />;
 };
 
-export const SnippetsV9 = (props?: any) => {
-  return <LazySnippetsV9 {...props} />;
-};
-
-export const CopyButtonV9 = (props?: any) => {
-  return <LazyCopyButtonV9 {...props} />;
+export const CopyButton = (props?: any) => {
+  return <LazyCopyButton {...props} />;
 };
 
 export const Auth = (props?: any) => {
-  return <LazyAuth {...props} />;
+  return LazyAuth ? <LazyAuth {...props} /> : null;
 };
 
 export const RequestHeaders = (props?: any) => {
