--- src/app/views/main-header/tenantIcon.tsx
@@ -1,15 +1,0 @@
-/* eslint-disable max-len */
-import { getTheme } from '@fluentui/react';
-
-const TenantIcon = () => {
-  const theme = getTheme();
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path d="M9.68034 7.5C9.83788 6.2359 10.091 5.11427 10.4095 4.20788C10.6874 3.41703 11.001 2.83044 11.3119 2.45615C11.6265 2.07734 11.8628 2 12 2C12.1372 2 12.3735 2.07734 12.6881 2.45615C12.9991 2.83044 13.3126 3.41703 13.5905 4.20788C13.909 5.11427 14.1621 6.2359 14.3197 7.5H9.68034ZM8.99431 3.71065C9.16879 3.21406 9.36744 2.75385 9.5909 2.34623C7.03372 3.10058 4.97829 5.02516 4.04455 7.5H8.1695C8.33611 6.06974 8.61973 4.77676 8.99431 3.71065ZM15.0057 3.71065C14.8312 3.21406 14.6326 2.75385 14.4091 2.34623C16.9663 3.10058 19.0217 5.02516 19.9554 7.5H15.8305C15.6639 6.06974 15.3803 4.77676 15.0057 3.71065ZM15.9587 9H20.3681C20.4548 9.48691 20.5 9.98817 20.5 10.5C20.5 11.0118 20.4548 11.5131 20.3681 12H15.9587C15.9859 11.5102 16 11.0092 16 10.5C16 9.99082 15.9859 9.48981 15.9587 9ZM15.8305 13.5H19.9554C19.6083 14.42 19.1062 15.2639 18.481 16C18.4755 16.0065 18.47 16.0129 18.4645 16.0194C17.9905 16.5739 17.4465 17.0668 16.846 17.4842C16.11 17.9958 15.2893 18.3941 14.4091 18.6538C14.6326 18.2461 14.8312 17.7859 15.0057 17.2894C15.3803 16.2232 15.6639 14.9303 15.8305 13.5ZM12.6881 18.5439C12.3735 18.9227 12.1372 19 12 19C11.8628 19 11.6265 18.9227 11.3119 18.5439C11.001 18.1696 10.6874 17.583 10.4095 16.7921C10.091 15.8857 9.83788 14.7641 9.68034 13.5H14.3197C14.1621 14.7641 13.909 15.8857 13.5905 16.7921C13.3126 17.583 12.9991 18.1696 12.6881 18.5439ZM7.15396 17.4842C7.88998 17.9958 8.71072 18.3941 9.5909 18.6538C9.36744 18.2461 9.16878 17.7859 8.99431 17.2894C8.61973 16.2232 8.33611 14.9303 8.1695 13.5H4.04455C4.39166 14.42 4.89376 15.2639 5.51903 16C5.52452 16.0065 5.53002 16.0129 5.53553 16.0194C6.00947 16.5739 6.55352 17.0668 7.15396 17.4842ZM8.04133 12H3.63193C3.54524 11.5131 3.5 11.0118 3.5 10.5C3.5 9.98817 3.54524 9.48691 3.63193 9H8.04133C8.01407 9.48981 8 9.99082 8 10.5C8 11.0092 8.01407 11.5102 8.04133 12ZM9.54378 9C9.51505 9.48646 9.5 9.98756 9.5 10.5C9.5 11.0124 9.51505 11.5135 9.54378 12H14.4562C14.4849 11.5135 14.5 11.0124 14.5 10.5C14.5 9.98756 14.4849 9.48646 14.4562 9H9.54378ZM4.00001 16H4.2531C4.64113 16.5456 5.08534 17.0484 5.5773 17.5H4.57007L3.73674 20.5H20.2633L19.4299 17.5H18.4227C18.9147 17.0484 19.3589 16.5456 19.7469 16H20C20.3369 16 20.6325 16.2247 20.7226 16.5493L21.9726 21.0493C22.0354 21.2751 21.9889 21.5173 21.8471 21.7038C21.7053 21.8904 21.4844 22 21.25 22H2.75001C2.51563 22 2.29474 21.8904 2.15291 21.7038C2.01108 21.5173 1.96464 21.2751 2.02737 21.0493L3.27737 16.5493C3.36754 16.2247 3.6631 16 4.00001 16Z"
-        fill={theme.palette.themePrimary} />
-    </svg>
-  )
-}
-
-export default TenantIcon;
-
