+++ src/app/views/query-response/headers/ResponseHeaders.tsx
@@ -1,40 +1,32 @@
-
 import { RESPONSE_HEADERS_COPY_BUTTON } from '../../../../telemetry/component-names';
 
-import { trackedGenericCopy } from '../../common/copy';
-import { CopyButtonV9 } from '../../common/lazy-loader/component-registry';
-import { convertVhToPx, getResponseEditorHeight,
-  getResponseHeight } from '../../common/dimensions/dimensions-adjustment';
 import { useAppSelector } from '../../../../store';
-import { MonacoV9 } from '../../common';
+import { Monaco } from '../../common';
+import { trackedGenericCopy } from '../../common/copy';
+import { CopyButton } from '../../common/lazy-loader/component-registry';
 
 const ResponseHeaders = () => {
-  const { dimensions: { response }, graphResponse, responseAreaExpanded, sampleQuery } =
-    useAppSelector((state) => state);
+  const graphResponse = useAppSelector((state) => state.graphResponse);
+
+  const sampleQuery = useAppSelector((state) => state.sampleQuery);
   const { headers } = graphResponse.response;
-
-  const defaultHeight = convertVhToPx(getResponseHeight(response.height, responseAreaExpanded), 220);
-  const monacoHeight = getResponseEditorHeight(120);
-
-
-  const handleCopy = async () => trackedGenericCopy(JSON.stringify(headers), RESPONSE_HEADERS_COPY_BUTTON, sampleQuery)
+  const handleCopy = async () =>
+    trackedGenericCopy(
+      JSON.stringify(headers),
+      RESPONSE_HEADERS_COPY_BUTTON,
+      sampleQuery
+    );
 
   if (headers) {
     return (
-      <div id='response-headers-tab'>
-        <CopyButtonV9
-          handleOnClick={handleCopy}
-          isIconButton={true}
-          style={{ float: 'right', zIndex: 1 }}
-        />
-        <MonacoV9 body={headers} height={responseAreaExpanded ? defaultHeight : monacoHeight} />
-      </div>
+      <>
+        <CopyButton handleOnClick={handleCopy} isIconButton={true} />
+        <Monaco body={headers} height='25rem' />
+      </>
     );
   }
 
-  return (
-    <div />
-  );
+  return <div />;
 };
 
 export default ResponseHeaders;
