+++ src/app/views/query-response/snippets/index.tsx
@@ -1,4 +1,3 @@
-// import Snippets from './Snippets';
-import SnippetsV9 from './SnippetsV9';
-export { SnippetsV9 };
+import Snippets from './Snippets';
+export default Snippets;
