+++ src/app/views/query-runner/query-input/auto-complete/suggestion-list/SuggestionsList.tsx
@@ -1,15 +1,9 @@
-import { Label, getTheme, mergeStyles } from '@fluentui/react';
+import { Label } from '@fluentui/react-components';
 import { createRef, useEffect } from 'react';
 
 import { ISuggestionsList } from '../../../../../../types/auto-complete';
-import { autoCompleteStyles } from '../AutoComplete.styles';
 
 const SuggestionsList = ({ filteredSuggestions, activeSuggestion, onSuggestionSelected }: ISuggestionsList) => {
-  const theme = getTheme();
-  const suggestionsClass = mergeStyles(autoCompleteStyles(theme).suggestions);
-  const suggestionActiveClass = mergeStyles(autoCompleteStyles(theme).suggestionActive);
-  const suggestionOptionClass = mergeStyles(autoCompleteStyles(theme).suggestionOption);
-  const suggestionTitleClass = mergeStyles(autoCompleteStyles(theme).suggestionTitle);
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const refs = filteredSuggestions.reduce((ref: any, value: string) => {
@@ -29,16 +23,15 @@ const SuggestionsList = ({ filteredSuggestions, activeSuggestion, onSuggestionSe
   }, [activeSuggestion]);
 
   return (
-    <ul className={suggestionsClass} tabIndex={-1}>
+    <ul tabIndex={-1}>
       {filteredSuggestions.map((suggestion: string, index: number) => {
         return (
           <li
-            className={(index === activeSuggestion) ? suggestionActiveClass : suggestionOptionClass}
             key={index}
             ref={refs[index]}
             onClick={() => onSuggestionSelected(suggestion)}
           >
-            <Label className={suggestionTitleClass}>
+            <Label>
               {suggestion}
             </Label>
           </li>
