+++ src/types/profile.ts
@@ -1,8 +1,6 @@
-import { ITheme } from '@fluentui/react';
 import { ACCOUNT_TYPE } from '../app/services/graph-constants';
 
 export interface IProfileProps {
-  theme?: ITheme;
   styles?: object;
   mobileScreen: boolean;
   actions?: {
