--- src/types/theme.ts
@@ -1,6 +1,0 @@
-import { ITheme } from '@fluentui/react';
-
-export interface IComponentStyleProps {
-  theme: ITheme;
-  className?: string;
-}
